Force user to set the destination of logs/metrics for jailer
User has to make an explicit choice, when running with jailer, regarding
the destination of logs and  metrics. This is done by passing either
'make-fifos' parameter to jailer, which will allow the creation of two fifos
(one for logs and one for metrics), or 'no-fifos' parameter (logs and metrics
stored only in stdout/stderr). We do not allow skipping these parameters
(one of them has to be present).

Signed-off-by: lauralg <lauralg@uc53015e8c5905e.ant.amazon.com>